requirements.txt
discord.py
python-dotenv
fuzzywuzzy
unidecode
openai